plugins/cmesh/tools/collab-sync/requirements.txt
pyyaml>=6.0
pyrage>=1.3.0
httpx>=0.27
